Made codacy happy - split tests out
